healer_service-main/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
rapidfuzz>=3.5.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pytest>=7.4.0
httpx>=0.25.0
streamlit 
sentence-transformers 
faiss-cpu 
scikit-learn
